Server Files_BAckup/Backup_13122022/multi_clients_MQTT/requirements.txt
neo4j==4.4.2
paho_mqtt==1.6.1
requests==2.25.1
